+++ app/com/lynxanalytics/biggraph/frontend_operations/Operations.scala
@@ -35,15 +35,16 @@ object OperationParams {
     val multipleChoice = false
     def validate(value: String): Unit = {}
   }
+
   case class Choice(
       id: String,
       title: String,
       options: List[FEOption],
       multipleChoice: Boolean = false,
-      allowUnknownOption: Boolean = false) extends OperationParameterMeta {
+      allowUnknownOption: Boolean = false,
+      mandatory: Boolean = true) extends OperationParameterMeta {
     val kind = "choice"
-    val defaultValue = ""
-    val mandatory = true
+    val defaultValue = options.headOption.map(_.id).getOrElse("")
     def validate(value: String): Unit = {
       if (!allowUnknownOption) {
         val possibleValues = options.map { x => x.id }.toSet
@@ -58,6 +59,7 @@ object OperationParams {
       }
     }
   }
+
   case class TableParam(
       id: String,
       title: String,
@@ -68,6 +70,7 @@ object OperationParams {
     val mandatory = true
     def validate(value: String): Unit = {}
   }
+
   case class TagList(
       id: String,
       title: String,
@@ -78,6 +81,7 @@ object OperationParams {
     val defaultValue = ""
     def validate(value: String): Unit = {}
   }
+
   case class File(id: String, title: String) extends OperationParameterMeta {
     val kind = "file"
     val multipleChoice = false
@@ -86,6 +90,7 @@ object OperationParams {
     val mandatory = true
     def validate(value: String): Unit = {}
   }
+
   case class Ratio(id: String, title: String, defaultValue: String = "")
       extends OperationParameterMeta {
     val kind = "default"
@@ -97,6 +102,7 @@ object OperationParams {
         s"$title ($value) has to be a ratio, a double between 0.0 and 1.0")
     }
   }
+
   case class NonNegInt(id: String, title: String, default: Int)
       extends OperationParameterMeta {
     val kind = "default"
@@ -108,6 +114,7 @@ object OperationParams {
       assert(value matches """\d+""", s"$title ($value) has to be a non negative integer")
     }
   }
+
   case class NonNegDouble(id: String, title: String, defaultValue: String = "")
       extends OperationParameterMeta {
     val kind = "default"
@@ -118,6 +125,7 @@ object OperationParams {
       assert(value matches """\d+(\.\d+)?""", s"$title ($value) has to be a non negative double")
     }
   }
+
   case class Code(
       id: String,
       title: String,
@@ -581,7 +589,7 @@ class Operations(env: SparkFreeEnvironment) extends OperationRepository(env) {
     def parameters = List(
       Param("name", "Segmentation name", defaultValue = "modular_clusters"),
       Choice("weights", "Weight attribute", options =
-        FEOption.noWeight +: edgeAttributes[Double]),
+        FEOption.noWeight +: edgeAttributes[Double], mandatory = false),
       Param(
         "max-iterations",
         "Maximum number of iterations to do",
@@ -593,7 +601,7 @@ class Operations(env: SparkFreeEnvironment) extends OperationRepository(env) {
     def enabled = hasEdgeBundle
     def apply(params: Map[String, String]) = {
       val edgeBundle = project.edgeBundle
-      val weightsName = params("weights")
+      val weightsName = params.getOrElse("weights", FEOption.noWeight.id)
       val weights =
         if (weightsName == FEOption.noWeight.id) edgeBundle.const(1.0)
         else project.edgeAttributes(weightsName).runtimeSafeCast[Double]
@@ -1106,15 +1114,16 @@ class Operations(env: SparkFreeEnvironment) extends OperationRepository(env) {
     def parameters = List(
       Param("name", "Attribute name", defaultValue = "page_rank"),
       Choice("weights", "Weight attribute",
-        options = FEOption.noWeight +: edgeAttributes[Double]),
+        options = FEOption.noWeight +: edgeAttributes[Double], mandatory = false),
       NonNegInt("iterations", "Number of iterations", default = 5),
       Ratio("damping", "Damping factor", defaultValue = "0.85"))
     def enabled = hasEdgeBundle
     def apply(params: Map[String, String]) = {
       assert(params("name").nonEmpty, "Please set an attribute name.")
       val op = graph_operations.PageRank(params("damping").toDouble, params("iterations").toInt)
+      val weightsName = params.getOrElse("weights", FEOption.noWeight.id)
       val weights =
-        if (params("weights") == FEOption.noWeight.id) project.edgeBundle.const(1.0)
+        if (weightsName == FEOption.noWeight.id) project.edgeBundle.const(1.0)
         else project.edgeAttributes(params("weights")).runtimeSafeCast[Double]
       project.newVertexAttribute(
         params("name"), op(op.es, project.edgeBundle)(op.weights, weights).result.pagerank, help)
@@ -2504,7 +2513,7 @@ class Operations(env: SparkFreeEnvironment) extends OperationRepository(env) {
       Choice("leftName", "First ID attribute", options = vertexAttributes[String]),
       Choice("rightName", "Second ID attribute", options = vertexAttributes[String]),
       Choice("weights", "Edge weights",
-        options = FEOption.noWeight +: edgeAttributes[Double]),
+        options = FEOption.noWeight +: edgeAttributes[Double], mandatory = false),
       NonNegInt("mo", "Minimum overlap", default = 1),
       Ratio("ms", "Minimum similarity", defaultValue = "0.5"),
       Param(
@@ -2521,8 +2530,9 @@ class Operations(env: SparkFreeEnvironment) extends OperationRepository(env) {
       assert(mo >= 1, "Minimum overlap cannot be less than 1.")
       val leftName = project.vertexAttributes(params("leftName")).runtimeSafeCast[String]
       val rightName = project.vertexAttributes(params("rightName")).runtimeSafeCast[String]
+      val weightsName = params.getOrElse("weights", FEOption.noWeight.id)
       val weights =
-        if (params("weights") == FEOption.noWeight.id) project.edgeBundle.const(1.0)
+        if (weightsName == FEOption.noWeight.id) project.edgeBundle.const(1.0)
         else project.edgeAttributes(params("weights")).runtimeSafeCast[Double]
 
       val candidates = {