+++ kitescripts/big_data_tests/centrality.groovy
@@ -1,0 +1,12 @@
+// Centrality benchmark.
+
+/// REQUIRE_SCRIPT edge_import.groovy
+
+project = lynx.loadProject('edge_import_result')
+
+project.centrality(algorithm: 'Harmonic', bits: '4', maxDiameter: '5', name: 'centrality')
+
+centrality_histogram = project.vertexAttributes['centrality'].histogram(
+  logarithmic: true,
+  precise: true)
+println "centrality: $centrality_histogram"