+++ tools/call_spark_submit.sh
@@ -55,7 +55,7 @@ addJPropIfNonEmpty https.keyStorePassword "${KITE_HTTPS_KEYSTORE_PWD}"
 addJPropIfNonEmpty application.secret "${KITE_APPLICATION_SECRET}"
 addJPropIfNonEmpty authentication.google.clientSecret "${KITE_GOOGLE_CLIENT_SECRET}"
 addJPropIfNonEmpty hadoop.tmp.dir "${KITE_LOCAL_TMP}"
-addJPropIfNonEmpty pidfile.path "${KITE_PID_FILE}"
+addJPropIfNonEmpty pidfile.path "/dev/null"
 addJPropIfNonEmpty http.netty.maxInitialLineLength 10000
 addJPropIfNonEmpty jdk.tls.ephemeralDHKeySize 2048
 
@@ -99,7 +99,7 @@ if [ "${SPARK_MASTER}" == "yarn-client" ]; then
   if [ -n "${YARN_EXECUTOR_MEMORY_OVERHEAD_MB}" ]; then
     COMPUTED_EXECUTOR_MEMORY_OVERHEAD_MB="${YARN_EXECUTOR_MEMORY_OVERHEAD_MB}"
   else
-    RATIO_PERCENT=15
+    RATIO_PERCENT=20
     LAST_CHAR=${EXECUTOR_MEMORY: -1}
     if [ "${LAST_CHAR}" == "m" ]; then
       COMPUTED_EXECUTOR_MEMORY_OVERHEAD_MB=$((${EXECUTOR_MEMORY::-1} * $RATIO_PERCENT / 100))
@@ -174,10 +174,6 @@ command=(
 )
 
 startKite () {
-  if [ -f "${KITE_PID_FILE}" ]; then
-    >&2 echo "LynxKite is already running (or delete ${KITE_PID_FILE})"
-    exit 1
-  fi
   if [ ! -d "${SPARK_HOME}" ]; then
     >&2 echo "Spark cannot be found at ${SPARK_HOME}"
     exit 1
@@ -298,5 +294,11 @@ case $mode in
     exit 1
   ;;
 esac
+
+if [ -n "${KITE_SCRIPT_LOGS}" ]; then
+    THIS_PROG="$(readlink -f "$0")"
+    NOW=`date "+%Y:%m:%d %H:%M:%S"`
+    echo $NOW $THIS_PROG $mode >> $KITE_SCRIPT_LOGS
+fi
 
 exit