+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Please add changes to "master", preferably ordered by their significance. (Most 
 
 ### master
 
+ - Prefix-based user access control can be specified in the prefix definitions file.
  - emr.sh: support setting up LynxKite in an Amazon VPC.
  - Numeric fields in CSV files can be imported to LynxKite with the right types.
  - When importing edges for existing vertices, you can now join by non-string attributes too.